requirements.txt
discord
python-dotenv
Flask
asgiref